+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+- Log a count of the files to be spellchecked.
 - Add a large number of unit and end-to-end tests.
 - Add development instructions to [README.md](./README.md).
 - Add support for the following [retext](https://github.com/retextjs/retext) plugins:
@@ -15,6 +16,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   - [`retext-repeated-words`](https://github.com/retextjs/retext-repeated-words)
   - [`retext-syntax-mentions`](https://github.com/retextjs/retext-syntax-mentions)
   - [`retext-syntax-urls`](https://github.com/retextjs/retext-syntax-urls)
+- Apply all of the above plugins, as well as [`retext-spell`](https://github.com/retextjs/retext-spell), by default.
+- Add a newline to the end of the generated dictionary.
 
 ## [2.0.0] - 2017-12-28